Separated pipeline into two files, one for a workspace and the other for functions. functions are still poorly annotated and partially redundant. Mostly I need to reduce the number of values I am passing to functions either by consolidating my data structures or by being smarter and taking advantage of variable scope. I should look up whether R passes by value or by reference... that could make a BIG difference for some of the new functions.
Added ability to scan through Z stacks for intensity within a defined area.
Added code to selectively study well-formed nucleii, that are round and are not clipped
by the border of the frame

I did a clever thing that uses the projection of all stacks to define areas to search
through other stacks. I should think hard to find a way to distinguish MS2 signal from
background. The easy way would be to identify a local noise distribution and call something
real signal if it is above that noise, but i'm wondering if there might be some way
to take advantage of the shape of a transcription focus. i.e. only accept points that
show a bivariate normal in x-y (or maybe in x-y-z if there is enugh data to impose that
constraint on the z axis). I think EBImage has a clever way to scan through an image
and only accept things as true points if they reflect a bivariate gaussian intensity
distribution, but i need to think a little bit more about whether that is even the
smart way to do. it is certainly better than pure thresholding but i am still very torn.
--> perhaps instead of calculating the INTENSITY within the defined focus area, calculate
    the likelihood that the focus in that plane is derived from a bivariate gaussian. That
    might make the background drop down a lot because it is effectively reading out
    coordinated structure in multiple pixels rather than sum of pixels with random value,
    which is guaranteed to have some (pretty broad) distribution that makes it tough to
    identify where the point actually is.